pyproject.toml
[tool.poetry]
name = "electricity-bill-alerter"
version = "0.1.0"
description = ""
authors = ["Mohsin-Ul-Islam <mohsinulislam180@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
requests = "^2.27.1"
twilio = "^7.6.0"
google-cloud-secret-manager = "^2.8.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
